refactor : 리팩토링으로 인한 변동
