+++ src/main/java/com/kakaotech/team14backend/oauth2/presentation/LoginController.java
@@ -4,7 +4,7 @@ import com.kakaotech.team14backend.auth.PrincipalDetails;
 import com.kakaotech.team14backend.common.ApiResponse;
 import com.kakaotech.team14backend.common.ApiResponseGenerator;
 import com.kakaotech.team14backend.common.MessageCode;
-import com.kakaotech.team14backend.exception.UserNotAuthenticatedException;
+import com.kakaotech.team14backend.member.exception.UserNotAuthenticatedException;
 import com.kakaotech.team14backend.oauth2.application.AuthenticationSuccessHandler;
 import com.kakaotech.team14backend.oauth2.application.LoginService;
 import com.kakaotech.team14backend.oauth2.application.LogoutService;
